django/requirements.txt
absl-py==1.3.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
construct==2.10.68
contourpy==1.0.5
cryptography==39.0.2
cv==1.0.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.1.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
environ==1.0
ephem==4.1.4
et-xmlfile==1.1.0
exceptiongroup==1.0.1
executing==1.0.0
fastjsonschema==2.16.2
filelock==3.12.2
flatbuffers==23.1.21
fonttools==4.37.3
frozendict==2.3.5
fsspec==2023.6.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gdown==4.5.3
gensim==4.3.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.8.0
html5lib==1.1
huggingface-hub==0.16.2
hyperlink==21.0.0
idna==3.3
image==1.5.33
incremental==21.3.0
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.11.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
labelImg==1.8.6
libclang==15.0.6.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
miniful==0.0.6
mistune==2.0.4
multitasking==0.0.11
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.8.1
notebook==6.4.12
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.1.0
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.8.0
ply==3.11
pomegranate==0.14.8
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.15.0
pyFUME==0.2.25
Pygments==2.13.0
PyJWT==2.7.0
PyMeeus==0.5.12
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
pysmi==0.3.4
pysnmp==4.4.12
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==304
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.1
scikit-surprise==1.1.3
scipy==1.9.1
seaborn==0.12.1
selenium==4.6.0
Send2Trash==1.8.0
setuptools-git==1.2
simpful==2.10.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.5.0
statsmodels==0.13.5
tensor==0.3.6
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.15.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.3
torch==1.12.1
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.30.2
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.23.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.15.0
wsproto==1.2.0
yfinance==0.2.12
zope.interface==5.4.0
